+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pagination of search results.
 - Handling OTAA devices registered on an external Join Server in the Console.
 - RxMetadata Location field from Gateway Server.
+- Channel mask encoding in LinkADR MAC command.
 
 ### Security